Inline eventhub processor class
No functional change, just bringing together all the related classes for
the eventhub streaming client.
